Extract dynamic code response construction
Move execution-result mapping, diagnostics, friendly guidance, exception details, and cancellation responses into a dedicated factory. Add synchronous characterization coverage without invoking the dynamic compile-and-run path.
